Clinic_site-Django/clinic/requirements.txt
argon2-cffi==20.1.0
asgiref==3.2.7
bcrypt==3.1.7
certifi==2020.4.5.2
cffi==1.14.0
chardet==3.0.4
conf==0.4.1
confusable-homoglyphs==3.2.0
defusedxml==0.6.0
Django==3.0.7
django-allauth==0.42.0
django-bootstrap-datepicker-plus==3.0.5
django-computed-property==0.3.0
django-config==0.1.4
django-crispy-forms==1.9.1
django-datetimepicker==3.14
django-multiselectfield==0.1.12
django-registration==2.4.1
django-tempus-dominus==5.1.2.13
django-tinymce==3.0.2
django-urls==1.1.2
django-widget-tweaks==1.4.8
django-widgets-improved==1.5.0
get==2019.4.13
idna==2.9
image==1.5.32
importlib-metadata==1.6.1
include==0.2.2
Jinja2==2.11.2
MarkupSafe==1.1.1
myapp==0.1.dev0
oauthlib==3.1.0
path==14.0.1
Pillow==7.1.2
post==2019.4.13
public==2019.4.13
pycparser==2.20
python3-openid==3.1.0
pytz==2020.1
query-string==2019.4.13
reportlab==3.5.42
request==2019.4.13
requests==2.23.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
zipp==3.1.0
